Renamed folders from PascalCase to angular's naming convention.
