+++ src/app/3D/material-animators/CreateAnimation.ts
@@ -1,8 +1,8 @@
-import { ProductItem } from "../models/ProductItem/ProductItem";
+import { ProductItem } from "../models/product-item/ProductItem";
 import { OnAnimationProgressCallback } from "./OnAnimationProgressCallback";
-import { CancelActiveEventCallback } from "../models/ProductItem/CancelActiveEventCallback";
-import { ActiveProductItemEventType } from "../models/ProductItem/ActiveProductItemEventType";
-import { ActiveProductItemEvent } from "../models/ProductItem/ActiveProductItemEvent";
+import { CancelActiveEventCallback } from "../models/product-item/CancelActiveEventCallback";
+import { ActiveProductItemEventType } from "../models/product-item/ActiveProductItemEventType";
+import { ActiveProductItemEvent } from "../models/product-item/ActiveProductItemEvent";
 
 // TODO: Try and refactor all this into a better system at some point (so never!) or maybe take a promise as input to continue the promise?
