+++ src/app/3D/material-animators/MaterialAnimationType.ts
