+++ src/app/3D/material-animators/OnAnimationProgressCallback.ts
