+++ src/app/3D/models/event-data/LoadingProgressEventData.ts
