+++ src/app/3D/models/event-data/ModelLoadedEventData.ts
