+++ src/app/3D/models/selectable-meshes-options/SelectableMeshesOption.ts
