+++ src/app/3D/models/selectable-meshes-options/SelectableObject3DUserData.ts
