+++ src/app/3D/models/selectable-meshes-options/SelectedOptionsType.ts
