+++ src/app/3D/models/selectable-meshes-options/SelectedSpecificColorsValue.ts
@@ -1,4 +1,4 @@
-import { MaterialAnimationType } from "../../MaterialAnimators/MaterialAnimationType";
+import { MaterialAnimationType } from "../../material-animators/MaterialAnimationType";
 
 export interface SelectedSpecificColorsValue {
   animationType: MaterialAnimationType;
