+++ src/app/3D/models/selectable-meshes-options/SelectedSpecificTexturesValue.ts
@@ -1,4 +1,4 @@
-import { MaterialAnimationType } from "../../MaterialAnimators/MaterialAnimationType";
+import { MaterialAnimationType } from "../../material-animators/MaterialAnimationType";
 import { SelectedSpecificTexture } from "./SelectedSpecificTexture";
 
 export interface SelectedSpecificTexturesValue {
