+++ src/app/sidebar/sidebar-free-color/sidebar-free-color.component.ts
@@ -7,7 +7,7 @@ import {
 } from "../../3D/utility/MaterialUtility";
 import { ProductConfiguratorService } from "../../product-configurator.service";
 import { clearEvents } from "../../3D/utility/ProductItemUtility";
-import { ActiveProductItemEventType } from "../../3D/models/ProductItem/ActiveProductItemEventType";
+import { ActiveProductItemEventType } from "../../3D/models/product-item/ActiveProductItemEventType";
 
 @Component({
   selector: "sidebar-free-color",
