+++ src/app/toolbar/toolbar-subitem-container/toolbar-subitem-container.component.ts
@@ -9,9 +9,9 @@ import {
   OnChanges,
   SimpleChanges
 } from "@angular/core";
-import { SubProductItem } from "../../3D/models/ProductItem/SubProductItem";
+import { SubProductItem } from "../../3D/models/product-item/SubProductItem";
 import { ProductConfiguratorService } from "../../product-configurator.service";
-import { ProductItem } from "../../3D/models/ProductItem/ProductItem";
+import { ProductItem } from "../../3D/models/product-item/ProductItem";
 import { throttle } from "../../utility/throttle-decorator";
 
 @Component({
